Drop obsolete code from the initial boilerplate
Related to 5b1a2b20de5abb79948297b062da8c2429c62aee
There are no references to the `generate_document.html` template in the codebase.
The same goes for the `INVITATION_VALIDITY_DURATION` setting, which arrived straigt from
https://github.com/suitenumerique/docs

WeasyPrint is (I believe) not used in the project, so it is a ghost dependency.